pyproject.toml
[build-system]
requires = ["setuptools>=67.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "advanced_sim"
version = "0.1.0"
description = "Simulador avançado de campeonatos de futebol com stats de jogadores e evolução de temporada."
authors = [
    { name="Vinicius", github="viniruggeri" }
]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
keywords = ["futebol", "simulação", "data-science", "stats", "futebol-sim"]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment",
    "Topic :: Scientific/Engineering :: Information Analysis"
]

# Dependências do projeto
dependencies = [
    "numpy>=1.27",
    "pandas>=2.1",
    "scipy>=1.11",
    "matplotlib>=3.8",
    "seaborn>=0.12",
    "tqdm>=4.70",
    "pyyaml>=6.0",
    "python-dateutil>=2.9",
    "requests>=2.31",
]

# Extras opcionais (visualização e análise mais avançada)
[project.optional-dependencies]
viz = ["plotly>=6.0", "dash>=3.3"]
ml = ["scikit-learn>=1.3", "xgboost>=1.7", "lightgbm>=4.4"]
web = ["fastapi>=0.111", "uvicorn>=0.26"]

[tool.setuptools.packages.find]
where = ["src"]
